fix(desktop-electron): le durcissement ne peut plus emporter le processus principal
Au lancement d'une lecture, Electron fermait l'application sur « A JavaScript
error occurred in the main process » : TypeError, cannot mix BigInt and other
types, dans harden().

Deux defauts, et le second est le vrai.

koffi rend un int64/uint64 en NUMBER tant que la valeur tient dans la plage
sure, en BigInt seulement au-dela. Un descripteur de fenetre et un mot de style
sont petits : c'est donc toujours un Number qui arrive, et les `as bigint` du
fichier etaient faux depuis l'origine. Ils ne se voyaient pas parce que le seul
code qui combinait ces valeurs -- harden() -- ne s'executait jamais. Le corriger
a donc revele du meme coup son erreur de type. Mesure :

  GetWindowLongPtrW()   -> number 0x96000000
  FindWindowExW (rien)  -> number 0
  style | BigInt(...)   -> TypeError
  bits(style) | 0x...n  -> ok
  SetWindowLongPtrW(hwnd BigInt, value BigInt) -> accepte

Mais surtout : ce code tourne dans des rappels de minuteur du processus
principal, ou une exception est FATALE. Une ligne fautive, et toute
l'application par terre en pleine lecture. Le durcissement et le calage sont des
conforts : leur echec degrade -- clics qui n'atteignent pas les controles, video
mal cadree -- il ne doit jamais empecher de regarder un film. C'est le contrat
que l'app Tauri enonce noir sur blanc (mpv_window.rs:137) et que la coquille
n'avait pas repris. sansFaillir() le pose sur les deux chemins de minuteur, et
laisse une trace en warn.

+++ apps/desktop-electron/src/main/video/mpv.ts
@@ -77,7 +77,11 @@ export function command(args: readonly string[]): string | null {
 function decodeProperty(format: number, data: unknown): unknown {
   if (!data || format === FORMAT.NONE) return null;
   if (format === FORMAT.FLAG) return (koffi.decode(data, "int") as number) !== 0;
-  if (format === FORMAT.INT64) return Number(koffi.decode(data, "int64") as bigint);
+  // Pas de `as bigint` ici : koffi rend un Number tant que la valeur tient dans
+  // la plage sûre. `Number()` accepte les deux, donc c'était sans conséquence —
+  // mais l'assertion était fausse, et la même a coûté un crash dans
+  // `videoWindow.ts` (voir `bits()`).
+  if (format === FORMAT.INT64) return Number(koffi.decode(data, "int64"));
   if (format === FORMAT.DOUBLE) return koffi.decode(data, "double") as number;
   if (format === FORMAT.STRING) {
     const ptr = koffi.decode(data, "void*") as unknown;
